TRQ-3440. Fix two issues around qrun interpreting the -L syntax.
Conflicts:
	src/server/node_manager.c

+++ src/include/machine.hpp
@@ -259,7 +259,7 @@ class Socket
     void displayAsJson(stringstream &out, bool include_jobs) const;
     void setId(int id);
     void addChip(); // used for unit tests
-    int  how_many_tasks_fit(const req &r, int place_type) const;
+    float how_many_tasks_fit(const req &r, int place_type) const;
     void place_all_execution_slots(req &r, allocation &task_alloc);
     bool spread_place(req &r, allocation &master, int execution_slots_per, int &remainder, bool chips);
     bool spread_place_pu(req &r, allocation &task_alloc, int &cores, int &lprocs, int &gpus, int &mics);
